+++ README.md
@@ -1,7 +1,7 @@
-# Listgp
+# listgpt
 
 ## Introduction
-Adapted from the Java mascot _Duke_, Listgp helps you keep track of tasks using a Command Line Interface, with an option to use a Graphical User Interface.
+Adapted from the Java mascot _Duke_, listgpt helps you keep track of tasks using a Command Line Interface, with an option to use a Graphical User Interface.
 
 > “Your mind is for having ideas, not holding them.” – David Allen [source](https://dansilvestre.com/productivity-quotes/)
