+++ src/main/java/listgpt/CLI.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import java.util.Scanner;
