+++ src/main/java/listgpt/Event.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
