+++ src/main/java/listgpt/InvalidTaskException.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 public class InvalidTaskException extends RuntimeException {
     public InvalidTaskException(String message) {
