+++ src/main/java/listgpt/Launcher.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import javafx.application.Application;
