+++ src/main/java/listgpt/Parser.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import java.time.DateTimeException;
 import java.util.regex.Matcher;
