+++ src/main/java/listgpt/Task.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 /**
  * Represents a task with a name and a completion state.
