+++ src/test/java/listgpt/DeadlineTest.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
