+++ src/test/java/listgpt/ParserTest.java
@@ -1,4 +1,4 @@
-package Listgp;
+package listgpt;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
