+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
    __________________________________
-   Hello! I'm Listgp!
+   Hello! I'm listgpt!
    What can I do for you?
    __________________________________
